refactoring navigations a bitch so i just remade the file and copied everything over
