+++ src/components/views/SingleEmployeeView.jsx
@@ -3,8 +3,8 @@ import "./Table.css"
 import "./styles/tailwindStyle.css"
 import { useSelector, useDispatch  } from "react-redux";
 import { useState } from 'react'
-import NavigationButtons from "../components/NavigationButtons";
 
+import NavigationButtons from "/NavigationButtons";
 
 export default function SingleEmployeeView({empl, dispactchType,
   formData,
